requirements.txt
asgiref==3.9.1
Django==5.2.4
sqlparse==0.5.3
django-browser-reload
pillow
tensorflow
keras==3.10.0
numpy